requirements.txt
setuptools<81
asgiref==3.10.0
Django==4.2.25
djangorestframework==3.16.1
django-filter==24.3
drf-spectacular==0.27.2
mysqlclient==2.2.7
PyMySQL==1.1.2
sqlparse==0.5.3
typing_extensions==4.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow
python-dotenv
gunicorn==21.2.0
whitenoise==6.6.0
openpyxl==3.1.2